requirements.txt
# Core dependencies
vanna[chromadb,openai,mysql]>=2.0.2
mysql-connector-python>=9.0.0
python-dotenv>=1.0.0
PyMySQL>=1.1.0

# CLI dependencies
click>=8.1.0
rich>=13.7.0
prompt_toolkit>=3.0.43

# Data handling
pandas>=2.1.0
tabulate>=0.9.0

# LLM
openai>=1.10.0

# Vector store
chromadb>=1.1.0

# SQL
sqlalchemy>=2.0.0
sqlparse>=0.4.4

# Development dependencies
pytest>=7.4.0
ruff>=0.4.0

# Optional
numpy>=1.26.0
